requirements.txt
opencv-python==4.8.1.78
Pillow==10.0.1